+++ auto_route/lib/src/auto_route_wrapper.dart
@@ -2,5 +2,5 @@ import 'package:flutter/material.dart';
 
 // clients will implement this class to provide a wrapped route.
 abstract class AutoRouteWrapper {
-  Widget get wrappedRoute;
+  Widget wrappedRoute(BuildContext context);
 }