+++ example/lib/screens/second_screen.dart
@@ -1,7 +1,9 @@
+import 'package:auto_route/auto_route.dart';
+import 'package:example/screens/nested_screens/nested_router.gr.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-class SecondScreen extends StatelessWidget {
+class SecondScreen extends StatelessWidget implements AutoRouteWrapper {
   final String message;
 
   SecondScreen({@required this.message});
@@ -13,14 +15,16 @@ class SecondScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return CupertinoTabScaffold(
-      tabBuilder: (ctx, index) => Center(child: tabViews[index]),
-      tabBar: CupertinoTabBar(
-        items: <BottomNavigationBarItem>[
-          BottomNavigationBarItem(icon: Icon(Icons.book)),
-          BottomNavigationBarItem(icon: Icon(Icons.notifications))
-        ],
+    return Scaffold(
+      appBar: AppBar(),
+      body: ExtendedNavigator<NestedRouter>(
+        router: NestedRouter(),
       ),
     );
+  }
+
+  @override
+  Widget wrappedRoute(BuildContext context) {
+    return this;
   }
 }